+++ NEWS.md
@@ -2,6 +2,7 @@
 
 - add option `plot` to `TheilSen`. `FALSE` can be useful when analysing data to extract the trend components and plot in otherways and when the `TheilSen` plot is not required.
 - add option `silent` to `TheilSen` to avoid printing updates to trend fitting. By default it is `FALSE`.
+- fix wrong ordering of names in `timeVariation` when more than one pollutant
 
 # openair 2.2-3